ai_agent.py
"""
services/ai_agent.py
====================
Sends player stats to OpenAI and returns a natural language prediction insight.

WHY THIS FILE EXISTS:
    The ML model gives us a number. The AI agent turns that number into a
    readable sentence that makes sense for the presentation and earns
    the "AI integration" points on your rubric. For example, instead of
    just showing "Predicted: 27.5 PTS", the AI generates:
    "LeBron has scored 28+ in 4 of his last 5 games and is trending upward.
    Expect a strong scoring performance tonight."

WHAT'S IN HERE:
    - AIAgent class     : Wraps the OpenAI API call
    - generate_insight(): Formats stats into a prompt and returns AI response
    - _build_prompt()   : Private method that structures the OpenAI prompt

ADVANCED PYTHON FEATURES USED (rubric: 5 pts):
    - Decorator pattern (@retry logic via try/except loop)
    - Type hints
    - f-string formatting
    - os.environ for safe API key loading
    - Exception handling with fallback behavior
"""

import os
import time
from typing import Optional

from openai import OpenAI
from dotenv import load_dotenv

from models.player import GameLog
from models.prediction import Prediction

# Load .env file so OPENAI_API_KEY is available
load_dotenv()


def retry(max_attempts: int = 3, delay: float = 2.0):
    """
    Decorator factory that retries a function up to max_attempts times
    if it raises an exception. Waits 'delay' seconds between attempts.

    This is the decorator pattern from your rubric.

    Usage:
        @retry(max_attempts=3, delay=2.0)
        def my_api_call():
            ...
    """
    def decorator(func):
        def wrapper(*args, **kwargs):
            last_error = None
            for attempt in range(1, max_attempts + 1):
                try:
                    return func(*args, **kwargs)
                except Exception as e:
                    last_error = e
                    if attempt < max_attempts:
                        time.sleep(delay)
            raise RuntimeError(
                f"Function '{func.__name__}' failed after {max_attempts} attempts. "
                f"Last error: {last_error}"
            )
        return wrapper
    return decorator


class AIAgent:
    """
    Generates natural language insights about a player's predicted performance
    using the OpenAI API.

    Usage:
        agent = AIAgent()
        insight = agent.generate_insight(game_log, prediction)
        print(insight)
        # "LeBron James has averaged 29.2 points over his last 10 games..."
    """

    def __init__(self, model: str = "gpt-3.5-turbo", max_tokens: int = 150):
        """
        Args:
            model:      OpenAI model name (from config.yaml)
            max_tokens: Maximum length of the AI response
        """
        self.model = model
        self.max_tokens = max_tokens
        self._client = OpenAI(api_key=os.environ.get("OPENAI_API_KEY"))

    @retry(max_attempts=3, delay=2.0)
    def generate_insight(self, game_log: GameLog, prediction: Prediction) -> str:
        """
        Sends the player's recent stats and prediction to OpenAI and returns
        a natural language insight string.

        The @retry decorator above means if the API call fails (e.g. network
        timeout), it will automatically try again up to 3 times.

        Args:
            game_log:   The player's recent game data
            prediction: The ML model's prediction result

        Returns:
            A 2-3 sentence insight string, or a fallback message if AI fails
        """
        prompt = self._build_prompt(game_log, prediction)

        response = self._client.chat.completions.create(
            model=self.model,
            max_tokens=self.max_tokens,
            temperature=0.7,
            messages=[
                {
                    "role": "system",
                    "content": (
                        "You are a concise NBA sports analyst. "
                        "Give 2-3 sentence insights about a player's upcoming performance "
                        "based on their recent stats. Be specific and use the numbers provided. "
                        "Do not make up stats not given to you."
                    )
                },
                {
                    "role": "user",
                    "content": prompt
                }
            ]
        )

        return response.choices[0].message.content.strip()

    def generate_fallback_insight(self, game_log: GameLog, prediction: Prediction) -> str:
        """
        Returns a basic insight without calling OpenAI.
        Used when the API key is not set or when testing offline.
        """
        avg_pts = game_log.average_points
        avg_ast = game_log.average_assists
        avg_reb = game_log.average_rebounds
        return (
            f"{game_log.player_name} has averaged {avg_pts:.1f} PTS, "
            f"{avg_ast:.1f} AST, and {avg_reb:.1f} REB over their last "
            f"{game_log.game_count()} games. "
            f"The model predicts {prediction.predicted_value:.1f} "
            f"{prediction.stat_category} next game "
            f"(Confidence: {prediction.confidence_label})."
        )

    # ── Private Methods ──────────────────────────────────────────────────────────

    def _build_prompt(self, game_log: GameLog, prediction: Prediction) -> str:
        """
        Formats the player's stats and prediction into a clear OpenAI prompt.
        Good prompt engineering = better AI responses.
        """
        recent_pts = [f"{g.points:.0f}" for g in game_log.games[:5]]
        recent_ast = [f"{g.assists:.0f}" for g in game_log.games[:5]]
        recent_reb = [f"{g.rebounds:.0f}" for g in game_log.games[:5]]

        return (
            f"Player: {game_log.player_name}\n"
            f"Last 5 games — Points: {', '.join(recent_pts)}\n"
            f"Last 5 games — Assists: {', '.join(recent_ast)}\n"
            f"Last 5 games — Rebounds: {', '.join(recent_reb)}\n"
            f"Season averages — {game_log.average_points:.1f} PTS, "
            f"{game_log.average_assists:.1f} AST, {game_log.average_rebounds:.1f} REB\n"
            f"ML Prediction for next game: {prediction.predicted_value:.1f} {prediction.stat_category} "
            f"(Model confidence: {prediction.confidence_label})\n\n"
            f"Provide a 2-3 sentence analyst insight about this player's upcoming performance."
        )
